+++ src/gwt/src/org/rstudio/core/client/theme/res/ThemeStyles.java
@@ -71,6 +71,8 @@ public interface ThemeStyles extends CssResource
    String toolbarButtonPushed();
    String emptyProjectMenu();
 
+   String menuRightImage();
+   
    String scrollableMenuBar();
 
    String moduleTabPanel();