+++ src/gwt/src/org/rstudio/studio/client/application/ui/ProjectPopupMenu.java
@@ -15,6 +15,7 @@
 package org.rstudio.studio.client.application.ui;
 
 import org.rstudio.core.client.theme.res.ThemeResources;
+import org.rstudio.core.client.theme.res.ThemeStyles;
 import org.rstudio.core.client.widget.ToolbarButton;
 import org.rstudio.core.client.widget.ToolbarPopupMenu;
 import org.rstudio.studio.client.RStudioGinjector;
@@ -25,8 +26,11 @@ import org.rstudio.studio.client.workbench.model.SessionInfo;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.dom.client.Document;
+import com.google.gwt.dom.client.Element;
 import com.google.gwt.resources.client.ClientBundle;
 import com.google.gwt.resources.client.ImageResource;
+import com.google.gwt.user.client.DOM;
+import com.google.gwt.user.client.Event;
 import com.google.inject.Inject;
 
 public class ProjectPopupMenu extends ToolbarPopupMenu
@@ -100,6 +104,40 @@ public class ProjectPopupMenu extends ToolbarPopupMenu
        
        return toolbarButton_;
    }
+   
+   @Override
+   protected ToolbarMenuBar createMenuBar()
+   {
+      return new SessionsPopupMenuBar();
+   }
+   
+   private class SessionsPopupMenuBar extends ToolbarMenuBar
+   {
+      public SessionsPopupMenuBar()
+      {
+         super(true);
+      }
+      
+      @Override
+      public void onBrowserEvent(Event event) {
+        Element element = DOM.eventGetTarget(event);
+        switch (DOM.eventGetType(event)) {
+          case Event.ONCLICK: {
+             if (element.getClassName().equals(ThemeStyles.INSTANCE.menuRightImage()))
+                ProjectMRUList.setOpenInNewWindow(true);
+          }
+        }
+        super.onBrowserEvent(event);
+      }
+   }
+   
+   @Override
+   public void getDynamicPopupMenu(DynamicPopupMenuCallback callback)
+   {
+      ProjectMRUList.setOpenInNewWindow(false);
+      callback.onPopupMenu(this);
+   }
+   
 
    interface Resources extends ClientBundle
    {
@@ -110,5 +148,5 @@ public class ProjectPopupMenu extends ToolbarPopupMenu
                               (Resources) GWT.create(Resources.class);
    private final String activeProjectFile_;
    private ToolbarButton toolbarButton_ = null;
-   private ProjectMRUList mruList_ ;
+   private ProjectMRUList mruList_;
 }